+++ Library/src/HTStyle.c
@@ -26,7 +26,7 @@
 **      -------------------
 */
 
-typedef float HTCoord;
+typedef double HTCoord;
 typedef int HTColor;
 
 typedef struct {
